[SPARK-57590][SQL] Add 7z archive coverage for Parquet
7z landed as a container (SPARK-58246), and Parquet's archive read is
container-agnostic (readLocalizedEntries -> localizeEntries -> readArchiveEntries
-> openArchiveStream, which now handles .7z), so Parquet reads 7z archives with
no production change. Adds ParquetSevenZArchiveReadSuite and, matching the
post-7z test layout on master, folds the per-container Parquet leaf suites into
ParquetArchiveReadBase (removing the standalone Tar/Zip suite files).
